Variavel numeroMagico, agora, é double, antes era int

+++ src/modulo-01-java-OO/Dia01/Orc.java
@@ -73,8 +73,8 @@ public class Orc
         }
     }
     
-    private int gerarNumero() {
-        int numeroMagico = 0;
+    private double gerarNumero() {
+        double numeroMagico = 0.0;
       
             // B
         if(this.vida >= 30 && this.vida <= 60) {
